requirements.txt
pyyaml
jinja2
ansible